Move physical device functions into their own module
