+++ vulkano/src/instance/mod.rs
@@ -111,27 +111,27 @@ pub use self::extensions::RawInstanceExtensions;
 pub use self::instance::ApplicationInfo;
 pub use self::instance::Instance;
 pub use self::instance::InstanceCreationError;
-pub use self::instance::MemoryHeap;
-pub use self::instance::MemoryHeapsIter;
-pub use self::instance::MemoryType;
-pub use self::instance::MemoryTypesIter;
-pub use self::instance::PhysicalDevice;
-pub use self::instance::PhysicalDeviceType;
-pub use self::instance::PhysicalDevicesIter;
-pub use self::instance::QueueFamiliesIter;
-pub use self::instance::QueueFamily;
 pub use self::layers::layers_list;
 pub use self::layers::LayerProperties;
 pub use self::layers::LayersIterator;
 pub use self::layers::LayersListError;
 pub use self::limits::Limits;
 pub use self::loader::LoadingError;
+pub use self::physical_device::MemoryHeap;
+pub use self::physical_device::MemoryHeapsIter;
+pub use self::physical_device::MemoryType;
+pub use self::physical_device::MemoryTypesIter;
+pub use self::physical_device::PhysicalDevice;
+pub use self::physical_device::PhysicalDeviceType;
+pub use self::physical_device::PhysicalDevicesIter;
+pub use self::physical_device::QueueFamiliesIter;
+pub use self::physical_device::QueueFamily;
 pub use crate::version::Version;
 
 pub mod debug;
-pub mod loader;
-
 mod extensions;
 mod instance;
 mod layers;
 mod limits;
+pub mod loader;
+mod physical_device;
